+++ google/cloud/aiplatform/preview/pipelinejobschedule/pipeline_job_schedules.py
@@ -15,7 +15,7 @@
 # limitations under the License.
 #
 
-from typing import Optional
+from typing import List, Optional
 
 from google.auth import credentials as auth_credentials
 from google.cloud.aiplatform import base
@@ -39,7 +39,7 @@ from google.cloud.aiplatform_v1.types import (
 from google.cloud.aiplatform_v1beta1.types import (
     pipeline_job as gca_pipeline_job_v1beta1,
 )
-
+from google.protobuf import field_mask_pb2 as field_mask
 
 _LOGGER = base.Logger(__name__)
 
@@ -52,7 +52,7 @@ _VALID_AR_URL = schedule_constants._VALID_AR_URL
 # Pattern for any JSON or YAML file over HTTPS.
 _VALID_HTTPS_URL = schedule_constants._VALID_HTTPS_URL
 
-_READ_MASK_FIELDS = schedule_constants._PIPELINE_JOB_SCHEDULE_READ_MASK_FIELDS
+_SCHEDULE_ERROR_STATES = schedule_constants._SCHEDULE_ERROR_STATES
 
 
 class PipelineJobSchedule(
@@ -68,6 +68,7 @@ class PipelineJobSchedule(
     ):
         """Retrieves a PipelineJobSchedule resource and instantiates its
         representation.
+
         Args:
             pipeline_job (PipelineJob):
                 Required. PipelineJob used to init the schedule.
@@ -255,3 +256,193 @@ class PipelineJobSchedule(
         )
 
         _LOGGER.info("View Schedule:\n%s" % self._dashboard_uri())
+
+    @classmethod
+    def list(
+        cls,
+        filter: Optional[str] = None,
+        order_by: Optional[str] = None,
+        project: Optional[str] = None,
+        location: Optional[str] = None,
+        credentials: Optional[auth_credentials.Credentials] = None,
+    ) -> List["PipelineJobSchedule"]:
+        """List all instances of this PipelineJobSchedule resource.
+
+        Example Usage:
+
+        aiplatform.PipelineJobSchedule.list(
+            filter='display_name="experiment_a27"',
+            order_by='create_time desc'
+        )
+
+        Args:
+            filter (str):
+                Optional. An expression for filtering the results of the request.
+                For field names both snake_case and camelCase are supported.
+            order_by (str):
+                Optional. A comma-separated list of fields to order by, sorted in
+                ascending order. Use "desc" after a field name for descending.
+                Supported fields: `display_name`, `create_time`, `update_time`
+            project (str):
+                Optional. Project to retrieve list from. If not set, project
+                set in aiplatform.init will be used.
+            location (str):
+                Optional. Location to retrieve list from. If not set, location
+                set in aiplatform.init will be used.
+            credentials (auth_credentials.Credentials):
+                Optional. Custom credentials to use to retrieve list. Overrides
+                credentials set in aiplatform.init.
+
+        Returns:
+            List[PipelineJobSchedule] - A list of PipelineJobSchedule resource objects.
+        """
+        return cls._list_with_local_order(
+            filter=filter,
+            order_by=order_by,
+            project=project,
+            location=location,
+            credentials=credentials,
+        )
+
+    def list_jobs(
+        self,
+        filter: Optional[str] = None,
+        order_by: Optional[str] = None,
+        enable_simple_view: bool = False,
+        project: Optional[str] = None,
+        location: Optional[str] = None,
+        credentials: Optional[auth_credentials.Credentials] = None,
+    ) -> List[PipelineJob]:
+        """List all PipelineJob 's created by this PipelineJobSchedule.
+
+        Example usage:
+
+        pipeline_job_schedule.list_jobs(order_by='create_time_desc')
+
+        Args:
+            filter (str):
+                Optional. An expression for filtering the results of the request.
+                For field names both snake_case and camelCase are supported.
+            order_by (str):
+                Optional. A comma-separated list of fields to order by, sorted in
+                ascending order. Use "desc" after a field name for descending.
+                Supported fields: `display_name`, `create_time`, `update_time`
+            enable_simple_view (bool):
+                Optional. Whether to pass the `read_mask` parameter to the list call.
+                Defaults to False if not provided. This will improve the performance of calling
+                list(). However, the returned PipelineJob list will not include all fields for
+                each PipelineJob. Setting this to True will exclude the following fields in your
+                response: `runtime_config`, `service_account`, `network`, and some subfields of
+                `pipeline_spec` and `job_detail`. The following fields will be included in
+                each PipelineJob resource in your response: `state`, `display_name`,
+                `pipeline_spec.pipeline_info`, `create_time`, `start_time`, `end_time`,
+                `update_time`, `labels`, `template_uri`, `template_metadata.version`,
+                `job_detail.pipeline_run_context`, `job_detail.pipeline_context`.
+            project (str):
+                Optional. Project to retrieve list from. If not set, project
+                set in aiplatform.init will be used.
+            location (str):
+                Optional. Location to retrieve list from. If not set, location
+                set in aiplatform.init will be used.
+            credentials (auth_credentials.Credentials):
+                Optional. Custom credentials to use to retrieve list. Overrides
+                credentials set in aiplatform.init.
+
+        Returns:
+            List[PipelineJob] - A list of PipelineJob resource objects.
+        """
+        list_filter = f"schedule_name={self._gca_resource.name}"
+        if filter:
+            list_filter = list_filter + f" AND {filter}"
+
+        return PipelineJob.list(
+            filter=list_filter,
+            order_by=order_by,
+            enable_simple_view=enable_simple_view,
+            project=project,
+            location=location,
+            credentials=credentials,
+        )
+
+    def update(
+        self,
+        display_name: Optional[str] = None,
+        cron_expression: Optional[str] = None,
+        start_time: Optional[str] = None,
+        end_time: Optional[str] = None,
+        allow_queueing: Optional[bool] = None,
+        max_run_count: Optional[int] = None,
+        max_concurrent_run_count: Optional[int] = None,
+    ) -> None:
+        """Update an existing PipelineJobSchedule.
+
+        Example usage:
+
+        pipeline_job_schedule.update(
+            display_name='updated-display-name',
+            cron_expression='1 2 3 4 5',
+        )
+
+        Args:
+            display_name (str):
+                Optional. The user-defined name of this PipelineJobSchedule.
+            cron_expression (str):
+                Optional. Time specification (cron schedule expression) to launch scheduled runs.
+                To explicitly set a timezone to the cron tab, apply a prefix: "CRON_TZ=${IANA_TIME_ZONE}" or "TZ=${IANA_TIME_ZONE}".
+                The ${IANA_TIME_ZONE} may only be a valid string from IANA time zone database.
+                For example, "CRON_TZ=America/New_York 1 * * * *", or "TZ=America/New_York 1 * * * *".
+            start_time (str):
+                Optional. Timestamp after which the first run can be scheduled.
+                If unspecified, it defaults to the schedule creation timestamp.
+            end_time (str):
+                Optional. Timestamp after which no more runs will be scheduled.
+                If unspecified, then runs will be scheduled indefinitely.
+            allow_queueing (bool):
+                Optional. Whether new scheduled runs can be queued when max_concurrent_runs limit is reached.
+            max_run_count (int):
+                Optional. Maximum run count of the schedule.
+                If specified, The schedule will be completed when either started_run_count >= max_run_count or when end_time is reached.
+            max_concurrent_run_count (int):
+                Optional. Maximum number of runs that can be started concurrently for this PipelineJobSchedule.
+
+        Raises:
+            RuntimeError: User tried to call update() before create().
+        """
+        pipeline_job_schedule = self._gca_resource
+        if pipeline_job_schedule.state in _SCHEDULE_ERROR_STATES:
+            raise RuntimeError(
+                "Not updating PipelineJobSchedule: PipelineJobSchedule must be active or completed."
+            )
+
+        updated_fields = []
+        if display_name is not None:
+            updated_fields.append("display_name")
+            setattr(pipeline_job_schedule, "display_name", display_name)
+        if cron_expression is not None:
+            updated_fields.append("cron")
+            setattr(pipeline_job_schedule, "cron", cron_expression)
+        if start_time is not None:
+            updated_fields.append("start_time")
+            setattr(pipeline_job_schedule, "start_time", start_time)
+        if end_time is not None:
+            updated_fields.append("end_time")
+            setattr(pipeline_job_schedule, "end_time", end_time)
+        if allow_queueing is not None:
+            updated_fields.append("allow_queueing")
+            setattr(pipeline_job_schedule, "allow_queueing", allow_queueing)
+        if max_run_count is not None:
+            updated_fields.append("max_run_count")
+            setattr(pipeline_job_schedule, "max_run_count", max_run_count)
+        if max_concurrent_run_count is not None:
+            updated_fields.append("max_concurrent_run_count")
+            setattr(
+                pipeline_job_schedule,
+                "max_concurrent_run_count",
+                max_concurrent_run_count,
+            )
+
+        update_mask = field_mask.FieldMask(paths=updated_fields)
+        self.api_client.update_schedule(
+            schedule=pipeline_job_schedule,
+            update_mask=update_mask,
+        )