+++ vertexai/_model_garden/_model_garden_models.py
@@ -15,9 +15,8 @@
 
 """Base class for working with Model Garden models."""
 
-import abc
 import dataclasses
-from typing import Dict, Optional, Type
+from typing import Dict, Optional, Type, TypeVar
 
 from google.cloud import aiplatform
 from google.cloud.aiplatform import base
@@ -25,6 +24,9 @@ from google.cloud.aiplatform import initializer as aiplatform_initializer
 from google.cloud.aiplatform import models as aiplatform_models
 from google.cloud.aiplatform import _publisher_models
 
+from google.cloud.aiplatform.compat.types import (
+    publisher_model as gca_publisher_model,
+)
 
 _SUPPORTED_PUBLISHERS = ["google"]
 
@@ -32,13 +34,24 @@ _SHORT_MODEL_ID_TO_TUNING_PIPELINE_MAP = {
     "text-bison": "https://us-kfp.pkg.dev/vertex-ai/large-language-model-pipelines/tune-large-model/sdk-1-25"
 }
 
+_SDK_PUBLIC_PREVIEW_LAUNCH_STAGE = frozenset(
+    [
+        gca_publisher_model.PublisherModel.LaunchStage.PUBLIC_PREVIEW,
+        gca_publisher_model.PublisherModel.LaunchStage.GA,
+    ]
+)
+_SDK_GA_LAUNCH_STAGE = frozenset([gca_publisher_model.PublisherModel.LaunchStage.GA])
+
 _LOGGER = base.Logger(__name__)
+
+T = TypeVar("T", bound="_ModelGardenModel")
 
 
 @dataclasses.dataclass
 class _ModelInfo:
     endpoint_name: str
     interface_class: Type["_ModelGardenModel"]
+    publisher_model_resource: _publisher_models._PublisherModel
     tuning_pipeline_uri: Optional[str] = None
     tuning_model_id: Optional[str] = None
 
@@ -107,11 +120,14 @@ def _get_model_info(
     )
 
     if not interface_class:
-        raise ValueError(f"Unknown model {publisher_model_res.name}")
+        raise ValueError(
+            f"Unknown model {publisher_model_res.name}; {schema_to_class_map}"
+        )
 
     return _ModelInfo(
         endpoint_name=endpoint_name,
         interface_class=interface_class,
+        publisher_model_resource=publisher_model_res,
         tuning_pipeline_uri=tuning_pipeline_uri,
         tuning_model_id=tuning_model_id,
     )
@@ -120,18 +136,30 @@ def _get_model_info(
 class _ModelGardenModel:
     """Base class for shared methods and properties across Model Garden models."""
 
-    @staticmethod
-    @abc.abstractmethod
-    def _get_public_preview_class_map() -> Dict[str, Type["_ModelGardenModel"]]:
-        """Returns a Dict mapping schema URI to model class.
+    _LAUNCH_STAGE: gca_publisher_model.PublisherModel.LaunchStage = (
+        _SDK_PUBLIC_PREVIEW_LAUNCH_STAGE
+    )
 
-        Subclasses should implement this method. Example mapping:
+    def _validate_launch_stage(
+        self,
+        publisher_model_resource: gca_publisher_model.PublisherModel,
+    ) -> None:
+        """Validates the model class _LAUNCH_STAGE matches the PublisherModel resource's launch stage.
 
-            {
-                "gs://google-cloud-aiplatform/schema/predict/instance/text_generation_1.0.0.yaml": TextGenerationModel
-            }
+        Args:
+            publisher_model_resource (gca_publisher_model.PublisherModel
+                The GAPIC PublisherModel resource for this model.
         """
-        pass
+
+        publisher_launch_stage = publisher_model_resource.launch_stage
+
+        if publisher_launch_stage not in self._LAUNCH_STAGE:
+            raise ValueError(
+                f"The model you are trying to instantiate does not support the launch stage: {publisher_launch_stage.name}"
+            )
+
+    # Subclasses override this attribute to specify their instance schema
+    _INSTANCE_SCHEMA_URI: Optional[str] = None
 
     def __init__(self, model_id: str, endpoint_name: Optional[str] = None):
         """Creates a _ModelGardenModel.
@@ -154,7 +182,7 @@ class _ModelGardenModel:
         )
 
     @classmethod
-    def from_pretrained(cls, model_name: str) -> "_ModelGardenModel":
+    def from_pretrained(cls: Type[T], model_name: str) -> T:
         """Loads a _ModelGardenModel.
 
         Args:
@@ -168,14 +196,21 @@ class _ModelGardenModel:
             ValueError: If model does not support this class.
         """
 
+        if not cls._INSTANCE_SCHEMA_URI:
+            raise ValueError(
+                f"Class {cls} is not a correct model interface class since it does not have an instance schema URI."
+            )
+
         model_info = _get_model_info(
-            model_id=model_name, schema_to_class_map=cls._get_public_preview_class_map()
+            model_id=model_name, schema_to_class_map={cls._INSTANCE_SCHEMA_URI: cls}
         )
 
         if not issubclass(model_info.interface_class, cls):
             raise ValueError(
                 f"{model_name} is of type {model_info.interface_class.__name__} not of type {cls.__name__}"
             )
+
+        cls._validate_launch_stage(cls, model_info.publisher_model_resource)
 
         return model_info.interface_class(
             model_id=model_name,